requirements.txt
requests==2.26.0
pytest==7.4.2
allure-pytest==2.13.2
